Se agrega otra figura geometrica al proyecto(Rectángulo)

+++ Herencia/FormasGeometricas/src/model/CalculosFormas.java
@@ -6,6 +6,6 @@ public interface CalculosFormas {
     
     public int calcularArea();
     
-    public double calcularPerimetro();
+    public int calcularPerimetro();
     
 }
